// packages/trading-core/src/tpsl/types/generated/tpsl-handler.ts
export enum PriceActionType {
  market = "market",
  last_close = "last_close",
}
export enum OrderType {
  NUMBER_0 = 0,
  NUMBER_1 = 1,
}
export enum ConditionalOrderType {
  stop_loss = "stop_loss",
  take_profit = "take_profit",
  send_quote = "send_quote",
}
export enum PositionType {
  NUMBER_0 = 0,
  NUMBER_1 = 1,
}
// ...
export enum ConditionalOrdersState {
  pending = "pending",
  new = "new",
  triggered = "triggered",
  triggered_pending = "triggered_pending",
  canceled = "canceled",
  killed = "killed",
}
export interface ConditionalOrderResponseSchema {
  quote_id: number;
  coh_quote_id: string;
  party_a_address: string;
  symbol_id: number;
  conditional_order_type: ConditionalOrderType;
  quantity: number;
  price?: number | null;
  conditional_order_price: number;
  order_type: OrderType;
  state: ConditionalOrdersState;
  action_price_type: PriceActionType;
  close_status?: string | null;
  position_type: PositionType;
  leverage?: number | string | null;
  create_time: number;
  modify_time: number;
}

export interface ConditionalOrderSearchRequestSchemaV3 {
  party_a_address?: string | null;
  symbol_id?: number | null;
  state?: ConditionalOrdersState[] | null;
  order_type?: OrderType | null;
  conditional_order_type?: ConditionalOrderType | null;
  conditional_price_type?: PriceActionType | null;
  start?: number | null;
  size?: number | null;
}